Added new feature
Previously I was working on something, but in between I have to add new
files for other collaborators due to their needs, so I have to stop
temporary stop what I'm doing and I'm not committing it.

Now I have done my part, I'll be committing and pushing to the server
now.